AI-Travel-Planner/requirements.txt
streamlit==1.30.0
langchain==0.1.0
langchain-google-genai==0.0.5
python-dotenv==1.0.0
pandas==2.1.0
plotly==5.18.0
